refactor(colorbar): move ColorBar spec into a shared cleopatra.colorbar module
Extract ColorBar, _resolve_colorbar, _swatch_text_default,
_DEPRECATED_CBAR_KWARGS, and _warn_deprecated_cbar_kwargs out of array_glyph.py
into a lean, glyph-independent cleopatra.colorbar module, so every glyph type can
import the typed spec without depending on array_glyph (issue #239). array_glyph
re-exports the names for backwards compatibility, so existing
'from cleopatra.array_glyph import ColorBar' imports keep working. No behaviour
change.
